references/manim-to-video-claw/storyboard/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
supabase==2.15.1
dotenv==0.9.9
json_repair==0.44.1
httpx==0.28.1
pydantic==2.11.4
PyYAML==6.0.2